requirements.txt
Pillow==1.7.8
pyexif==0.2.1
